fix: 헤더에서 fetchUser 수행하도록 수정
